refactor(user package): change package name from users to user and refactor folders
