+++ src/main/java/com/project/pillpal/user/dtos/JwtResponse.java
@@ -1,4 +1,4 @@
-package users.dtos;
+package com.project.pillpal.user.dtos;
 
 public record JwtResponse(String token) {
 }
