+++ src/main/java/com/project/pillpal/user/entity/Role.java
@@ -1,4 +1,4 @@
-package users;
+package com.project.pillpal.user.entity;
 
 public enum Role {
     ROLE_ADMIN,
