+++ src/main/java/com/project/pillpal/user/repository/UserRepository.java
@@ -1,5 +1,6 @@
-package users;
+package com.project.pillpal.user.repository;
 
+import com.project.pillpal.user.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface UserRepository extends JpaRepository<User, Long> {
